# swagger/customer-api.yaml
openapi: 3.0.3
info:
  title: Qmart Customer API
  description: |
    # Qmart Customer API Documentation

    Complete API documentation for Qmart fintech platform customer operations.

    ## Features
    - **Secure Authentication** - JWT-based authentication with OTP verification
    - **Digital Wallet** - Complete wallet management with transfers and QR codes
    - **KYC System** - 3-tier verification system with daily limits
    - **Transaction History** - Complete audit trail of all operations
    - **Security** - Rate limiting, input validation, and comprehensive error handling

    ## Authentication
    All protected endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Rate Limits
    - **Authentication**: 5 requests per 15 minutes
    - **Transfers**: 10 requests per minute
    - **General API**: 100 requests per 15 minutes
    - **KYC Submissions**: 3 requests per hour

    ## KYC Tier System
    - **Tier 1** (Default): ₦50,000 daily limit
    - **Tier 2** (NIN + BVN): ₦500,000 daily limit
    - **Tier 3** (Full verification): ₦5,000,000 daily limit

  version: 1.0.0
  contact:
    name: Qmart API Support
    email: support@qmart.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.qmart.com
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    KYCSubmissionRequest:
      type: object
      required: [tier, documents]
      properties:
        tier:
          type: integer
          enum: [2, 3]
          example: 2
          description: Requested KYC tier (2 or 3)
        documents:
          type: object
          properties:
            nin:
              type: string
              pattern: '^[0-9]{11}$'
              example: 12345678901
              description: 11-digit National Identification Number
            bvn:
              type: string
              pattern: '^[0-9]{11}$'
              example: 09876543210
              description: 11-digit Bank Verification Number
            selfieUrl:
              type: string
              format: uri
              example: https://storage.example.com/selfie.jpg
              description: URL to uploaded selfie image
            addressProofUrl:
              type: string
              format: uri
              example: https://storage.example.com/address.pdf
              description: URL to address verification document

    # Response Schemas
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          description: Response data (varies by endpoint)
...
paths:
...
  /api/kyc/submit:
    post:
      tags: [KYC Management]
      summary: Submit KYC documents
      description: |
        Submit KYC documents for tier upgrade.

        **Rate Limit:** 3 requests per hour
        **Authentication:** Required

        **Tier Requirements:**
        - **Tier 2**: NIN + BVN
        - **Tier 3**: NIN + BVN + Selfie + Address Proof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KYCSubmissionRequest'
      responses:
        '201':
          description: KYC submission successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          kycId:
                            type: string
                            example: 507f1f77bcf86cd799439012
                          status:
                            type: string
                            example: pending
                          tier:
                            type: integer
                            example: 2
                          submittedAt:
                            type: string
                            format: date-time
...
  /api/kyc/requirements/{tier}:
    get:
      tags: [KYC Management]
      summary: Get KYC tier requirements
      description: |
        Get requirements for specific KYC tier.

        **Rate Limit:** 50 requests per minute
        **Authentication:** Required
      parameters:
        - name: tier
          in: path
          required: true
          description: KYC tier (2 or 3)
          schema:
            type: integer
            enum: [2, 3]
      responses:
        '200':
          description: KYC requirements retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tier:
                            type: integer
                            example: 2
                          dailyLimit:
                            type: number
                            example: 500000
                          formattedDailyLimit:
                            type: string
                            example: ₦500,000
                          requirements:
                            type: array
                            items:
                              type: object
                              properties:
                                field:
                                  type: string
                                  example: nin
                                label:
                                  type: string
                                  example: National Identification Number
                                required:
                                  type: boolean
                                  example: true
                                format:
                                  type: string
                                  example: 11 digits